+++ docs/components/FormContainer.js
@@ -1,6 +1,7 @@
 import React from "react";
 
 import {
+  CentsField,
   EmailField,
   Form,
   NumberField,
@@ -8,12 +9,14 @@ import {
   StringField
 } from "../../src";
 
-const onSubmit = () => (
-  new Promise(resolve => setTimeout(() => resolve(), 1000))
-);
+const onSubmit = values => {
+  console.log(values); // eslint-disable-line no-console
+  return new Promise(resolve => setTimeout(() => resolve(), 1000));
+};
 
 const FormContainer = () => (
   <Form onSubmit={onSubmit}>
+    <CentsField label="Cents" name="cents" required />
     <EmailField label="Email" name="email" required />
     <NumberField label="Number" name="number" required />
     <PasswordField label="Password" name="password" required />