+++ src/components/Button.js
@@ -20,8 +20,8 @@ const Button = ({
   loading,
   ...props
 }) => (
-  <button
-    type={type} /* eslint react/button-has-type: "off" */
+  <button // eslint-disable-line react/button-has-type
+    type={type}
     className={
       classnames("chq-btn", className, {
         "chq-btn-iv": inverted,