+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - The `Panel` component.
 - The `EmailField`, `NumberField`, `StringField`, and `PasswordField` components.
 - The `Form` component.
+- The `CentsField` component.
 
 ### Changed
 - Renamed animations to be consistent with naming style throughout the rest of the component library e.g., `chqSpin`.